fix: migrate generate-ogp script from .mjs to .mts and update package.json
